+++ HydrationTimeStampContext.test.mjs
@@ -1,7 +1,6 @@
 import { strictEqual } from "assert";
 import React from "react";
 import ReactTestRenderer from "react-test-renderer";
-import { jsx } from "react/jsx-runtime.js";
 import HydrationTimeStampContext from "./HydrationTimeStampContext.mjs";
 import assertBundleSize from "./test/assertBundleSize.mjs";
 
@@ -17,10 +16,11 @@ export default (tests) => {
     const TestComponent = () => React.useContext(HydrationTimeStampContext);
     const contextValue = "a";
     const testRenderer = ReactTestRenderer.create(
-      jsx(HydrationTimeStampContext.Provider, {
-        value: contextValue,
-        children: jsx(TestComponent, {}),
-      })
+      React.createElement(
+        HydrationTimeStampContext.Provider,
+        { value: contextValue },
+        React.createElement(TestComponent)
+      )
     );
 
     strictEqual(testRenderer.toJSON(), contextValue);