+++ useCache.test.mjs
@@ -4,7 +4,7 @@ import {
   renderHook,
   suppressErrorOutput,
 } from "@testing-library/react-hooks/lib/pure.js";
-import { jsx } from "react/jsx-runtime.js";
+import React from "react";
 import Cache from "./Cache.mjs";
 import CacheContext from "./CacheContext.mjs";
 import assertBundleSize from "./test/assertBundleSize.mjs";
@@ -36,10 +36,7 @@ export default (tests) => {
     () => {
       try {
         const wrapper = ({ children }) =>
-          jsx(CacheContext.Provider, {
-            value: true,
-            children,
-          });
+          React.createElement(CacheContext.Provider, { value: true }, children);
 
         const revertConsole = suppressErrorOutput();
 
@@ -62,10 +59,7 @@ export default (tests) => {
   tests.add("`useCache` getting the cache.", () => {
     try {
       const wrapper = ({ cache, children }) =>
-        jsx(CacheContext.Provider, {
-          value: cache,
-          children,
-        });
+        React.createElement(CacheContext.Provider, { value: cache }, children);
 
       const cacheA = new Cache();