+++ useLoadOnStale.test.mjs
@@ -4,7 +4,7 @@ import {
   renderHook,
   suppressErrorOutput,
 } from "@testing-library/react-hooks/lib/pure.js";
-import { jsx } from "react/jsx-runtime.js";
+import React from "react";
 import revertableGlobals from "revertable-globals";
 import Cache from "./Cache.mjs";
 import CacheContext from "./CacheContext.mjs";
@@ -85,10 +85,7 @@ export default (tests) => {
     () => {
       try {
         const wrapper = ({ children }) =>
-          jsx(CacheContext.Provider, {
-            value: true,
-            children,
-          });
+          React.createElement(CacheContext.Provider, { value: true }, children);
 
         const revertConsole = suppressErrorOutput();
 
@@ -142,10 +139,7 @@ export default (tests) => {
     }
 
     const wrapper = ({ cache, children }) =>
-      jsx(CacheContext.Provider, {
-        value: cache,
-        children,
-      });
+      React.createElement(CacheContext.Provider, { value: cache }, children);
 
     try {
       const { result, rerender } = renderHook(