+++ useLoading.test.mjs
@@ -4,7 +4,7 @@ import {
   renderHook,
   suppressErrorOutput,
 } from "@testing-library/react-hooks/lib/pure.js";
-import { jsx } from "react/jsx-runtime.js";
+import React from "react";
 import Loading from "./Loading.mjs";
 import LoadingContext from "./LoadingContext.mjs";
 import assertBundleSize from "./test/assertBundleSize.mjs";
@@ -36,10 +36,11 @@ export default (tests) => {
     () => {
       try {
         const wrapper = ({ children }) =>
-          jsx(LoadingContext.Provider, {
-            value: true,
-            children,
-          });
+          React.createElement(
+            LoadingContext.Provider,
+            { value: true },
+            children
+          );
 
         const revertConsole = suppressErrorOutput();
 
@@ -62,10 +63,11 @@ export default (tests) => {
   tests.add("`useLoading` getting the loading.", () => {
     try {
       const wrapper = ({ loading, children }) =>
-        jsx(LoadingContext.Provider, {
-          value: loading,
-          children,
-        });
+        React.createElement(
+          LoadingContext.Provider,
+          { value: loading },
+          children
+        );
 
       const loadingA = new Loading();