+++ useLoadingEntry.test.mjs
@@ -5,7 +5,7 @@ import {
   renderHook,
   suppressErrorOutput,
 } from "@testing-library/react-hooks/lib/pure.js";
-import { jsx } from "react/jsx-runtime.js";
+import React from "react";
 import revertableGlobals from "revertable-globals";
 import Cache from "./Cache.mjs";
 import Loading from "./Loading.mjs";
@@ -65,10 +65,11 @@ export default (tests) => {
     () => {
       try {
         const wrapper = ({ children }) =>
-          jsx(LoadingContext.Provider, {
-            value: true,
-            children,
-          });
+          React.createElement(
+            LoadingContext.Provider,
+            { value: true },
+            children
+          );
 
         const revertConsole = suppressErrorOutput();
 
@@ -95,10 +96,11 @@ export default (tests) => {
         const loading = new Loading();
         const cache = new Cache();
         const wrapper = ({ children }) =>
-          jsx(LoadingContext.Provider, {
-            value: loading,
-            children,
-          });
+          React.createElement(
+            LoadingContext.Provider,
+            { value: loading },
+            children
+          );
 
         const cacheKeyA = "a";
 
@@ -213,10 +215,11 @@ export default (tests) => {
         );
 
         const wrapper = ({ children }) =>
-          jsx(LoadingContext.Provider, {
-            value: loading,
-            children,
-          });
+          React.createElement(
+            LoadingContext.Provider,
+            { value: loading },
+            children
+          );
 
         const { result, rerender } = renderHook(
           ({ cacheKey }) => useLoadingEntry(cacheKey),